Add some new path config to storage
