+++ babel.config.js
@@ -2,5 +2,16 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
+    plugins: [
+      [
+        'module-resolver',
+        {
+          alias: {
+            'storage': './src/storage',
+            'storage/*': './src/storage/*'
+          }
+        }
+      ]
+    ]
   };
 };
