+++ src/storage/package.json
@@ -1,0 +1,3 @@
+{
+    "name": "storage"
+}
