+++ tsconfig.json
@@ -7,8 +7,13 @@
     "noEmit": true,
     "skipLibCheck": true,
     "resolveJsonModule": true,
-    "strict": true
+    "strict": true,
+    "baseUrl": ".",
+    "paths": {
+      "storage": ["src/storage"],
+      "storage/*": ["src/storage/*"],
+    }
   },
   "include": ["App.tsx", "src/**/*"],
-  "exclude": ["node_modules", "**/*.spec.ts"]
+  "exclude": ["node_modules", "**/*.spec.ts"],
 }
